refactor: introduce layer between compiler and parser
