+++ prql-compiler/src/parser/prql_parser/mod.rs
@@ -3,8 +3,6 @@ mod interpolation;
 mod lexer;
 mod stmt;
 
-use std::collections::HashMap;
-
 use anyhow::Result;
 use chumsky::{
     error::{Cheap, SimpleReason},
@@ -15,13 +13,10 @@ use itertools::Itertools;
 
 use self::lexer::Token;
 
-use super::ast::pl::*;
+use crate::ast::pl::*;
 
-use crate::{
-    error::{Error, Errors, Reason, WithErrorInfo},
-    utils::IdGenerator,
-};
-use crate::{SourceTree, Span};
+use crate::error::{Error, Errors, Reason, WithErrorInfo};
+use crate::Span;
 
 mod span;
 
@@ -56,25 +51,6 @@ pub fn parse_source(source: &str, source_id: u16) -> Result<Vec<Stmt>> {
     } else {
         Err(Errors(errors).into())
     }
-}
-
-pub fn parse(file_tree: &SourceTree<String>) -> Result<SourceTree<Vec<Stmt>>> {
-    let mut res = SourceTree::default();
-
-    let ids: HashMap<_, _> = file_tree.source_ids.iter().map(|(a, b)| (b, a)).collect();
-    let mut id_gen = IdGenerator::<usize>::new();
-
-    for (path, source) in &file_tree.sources {
-        let id = ids
-            .get(path)
-            .map(|x| **x)
-            .unwrap_or_else(|| id_gen.gen() as u16);
-        let stmts = parse_source(source, id)?;
-
-        res.sources.insert(path.clone(), stmts);
-        res.source_ids.insert(id, path.clone());
-    }
-    Ok(res)
 }
 
 /// Helper that does not track source_ids
